+++ CHANGELOG.md
@@ -9,7 +9,8 @@ Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and
 - update `DiscordChatExporter` to v2.43.3
 - update and pinned `openai` to v1.35.5
 - update reduction method to use an offset, a selection mode of either sequential or distributed, and a flag to iterate in reverse to reduce the line selection
-- update some default values
+- update some default CLI command argument values
+- update some CLI help descriptions
 - major refactors
 
 ## [2.0.1] - 06-15-2023